Implement packages in the project
Unfortunately, this has to be done despite the troubles we had with our IDEs, we have to submit the project as a JAR and to do so, we need packages. We should do this earlier on to troubleshoot any issues earlier rather than later. Otherwise, we may have to reverse many changes to fix this later as compared to any earlier issues.